back-end/requirements.txt
Flask
werkzeug
email_validator
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
gunicorn
requests
python-dotenv
smtplib
schedule
PyMuPDF
python-docx
python-dotenv
pyodbc
azure-storage-blob
pymssql